requirements.txt
streamlit
google-generativeai
duckduckgo_search